Move lib/mastodon/redis_config for zeitwerk
