+++ config/redis.rb
